Split CLI code from library code
Previously the click handlers and the nanogen "library" code were all
combined into a single file. This is a little messy and makes it all
harder to understand, so I split out the CLI code to be its own module
that simply uses functions from the nanogen module.

There's some some weird coupling in the `preview` function right now,
but I'll solve it later.

Also updated some broken imports that were probably causing test
failures and a generally not-working-at-all bit of code.