Put the `nano` back in `nanogen`
Removed a lot of the customizability (no more YAML front-matter or
site.config) in favor of a simpler and less-flexible system. Also
reorganized the code to move away from a packaged-based layout toward
a looser collection of modules. Changes to documentation and the
included example were also made to support the re-org and
simplification.